bank

+++ banking-app/components/account/service/account.js
@@ -7,11 +7,14 @@ const Bank = require('../../../view/bank');
 // Take user class of index.js from account folder.
 const User = require('../../../view/user');
 
+// Take user class of index.js from account folder.
+const Transaction = require('../../../view/transaction');
+
 // Take exports of index.js from error folder.
 const CustomError = require('../../../errors')
 
 // Take db of index.js from models folder.
-const db = require("../../../models/transaction");
+const db = require("../../../models/index");
 
 // Create function for get all accounts.
 const getAllAccounts = async () => {
@@ -24,9 +27,9 @@ const getAllAccounts = async () => {
         // Create bucket for storing all accounts after getting it from view.
         const allAccounts = await Account.getAllAccounts(transaction)
 
-        // If no accounts found then send error.
-        if (!allAccounts || (allAccounts && allAccounts.length == 0)) {
-            throw new CustomError.BadRequestError("Accounts not found")
+        // If no accounts found then send empty array.
+        if (!allAccounts) {
+            allAccounts = []
         }
 
         // Commit the transaction.
@@ -99,9 +102,15 @@ const createAccount = async (accountObj) => {
             throw new CustomError.BadRequestError("User not found")
         }
 
+        // Check if created by exists in db.
+        const loggedInUser = User.createBlankUserWithID(accountObj.createdBy)
+        const loggedInUserForExists = await loggedInUser.getUserByID(transaction)
+        if (!loggedInUserForExists) {
+            throw new CustomError.BadRequestError("User not found")
+        }
+
         // For loop till account number is unique.
         createAccountNumberLabel: {
-            // for (let i = 0; ; i++) {
 
             // Create a random number for account number.
             let accountNumber = Math.floor((Math.random() * 10000) + 1)
@@ -114,14 +123,43 @@ const createAccount = async (accountObj) => {
             if (account) {
                 break createAccountNumberLabel
             }
-            // if (!account) {
-            //     break
-            // }
-            // }
         }
+
+        // Add the balance of the account to the total balance of user.
+        const userForUpdate = User.createUserWithID(userForExists.id, userForExists.firstName,
+            userForExists.lastName, userForExists.email, +userForExists.totalBalance + accountObj.balance,
+            userForExists.isAdmin, userForExists.password)
+
+        // Give updated by to user.
+        userForUpdate.updatedBy = accountObj.createdBy
+
+        // Update the user.
+        await userForUpdate.updateUser(transaction)
 
         // Create bucket for storing new account after getting it from view.
         let newAccount = await accountObj.createAccount(transaction)
+
+        // Create transaction bucket.
+        let transactionBucket = new Transaction("Deposit", accountObj.id, null, accountObj.balance)
+
+        // Give closing balance to transaction.
+        transactionBucket.closingBalance = accountObj.balance
+
+        // Give created by to transaction.
+        transactionBucket.createdBy = accountObj.createdBy
+
+        // Create the transaction.
+        await transactionBucket.createTransaction(transaction)
+
+        // Add the balance of the account to the total balance of bank.
+        const bankForUpdate = Bank.createBankWithID(bankForExists.id, bankForExists.name,
+            bankForExists.abbrevieation, +bankForExists.balance + accountObj.balance)
+
+        // Give updated by to bank.
+        bankForUpdate.updatedBy = accountObj.createdBy
+
+        // Update the user.
+        await bankForUpdate.updateBank(transaction)
 
         // Commit the transaction.
         await transaction.commit()
@@ -161,6 +199,13 @@ const updateAccount = async (accountObj) => {
             throw new CustomError.BadRequestError("User not found")
         }
 
+        // Check if updated by exists in db.
+        const checkUser = User.createBlankUserWithID(bankObj.updatedBy)
+        const loggedInuUserForExists = await checkUser.getUserByID(transaction)
+        if (!loggedInuUserForExists) {
+            throw new CustomError.BadRequestError("User not found")
+        }
+
         // Create bucket for storing updated account after getting it from view.
         let updateAccount = await accountObj.updateAccount(transaction)
 
@@ -187,6 +232,13 @@ const deleteAccount = async (accountObj) => {
     const transaction = await db.sequelize.transaction()
 
     try {
+
+        // Check if deleted by exists in db.
+        const checkUser = User.createBlankUserWithID(bankObj.deletedBy)
+        const userForExists = await checkUser.getUserByID(transaction)
+        if (!userForExists) {
+            throw new CustomError.BadRequestError("User not found")
+        }
 
         // Create bucket for storing deleted account after getting it from view.
         let deleteAccount = await accountObj.deleteAccount(transaction)
