+++ banking-app/components/user/service/user.js
@@ -2,46 +2,35 @@
 const User = require('../../../view/user');
 
 // Take db of index.js from models folder.
-const db = require("../../../models/transaction")
+const db = require("../../../models")
 
 // Import custom error.
 const CustomError = require('../../../errors')
 
 // Create function for get all users.
-const getAllUsers = async () => {
+const getAllUsers = async (params) => {
 
     // Start new transaction.
     const transaction = await db.sequelize.transaction()
 
     try {
 
-        // Create bucket for storing all users after getting it from view.
-        const allUsers = await User.getAllUsers(transaction)
+        // Format the search queries.
+        const searchQueries = addSearchQueries(params)
 
-        // If no users found then send error.
-        if (!allUsers || (allUsers && allUsers.length == 0)) {
-            throw new CustomError.BadRequestError("Users not found")
+        // Create bucket for storing all users after getting it from view.
+        const allUsers = await User.getAllUsers(transaction, searchQueries)
+
+        // If no users found then send empty array.
+        if (!allUsers) {
+            allUsers = []
         }
 
         // Commit the transaction.
         await transaction.commit()
 
-        // Format the allUsers bucket.
-        let allUsersFormattedList = []
-        for (let i = 0; i < allUsers.length; i++) {
-            allUsersFormattedList.push({
-                id: allUsers[i].id,
-                firstName: allUsers[i].firstName,
-                lastName: allUsers[i].lastName,
-                email: allUsers[i].email,
-                isAdmin: allUsers[i].isAdmin,
-                totalBalance: allUsers[i].totalBalance,
-                password: allUsers[i].password
-            })
-        }
-
         // Return bucket.
-        return allUsersFormattedList
+        return allUsers
 
     } catch (error) {
 
@@ -99,6 +88,13 @@ const createUser = async (userObj) => {
             throw new CustomError.BadRequestError("User email already exists")
         }
 
+        // Check if created by exists in db.
+        const checkUser = User.createBlankUserWithID(userObj.createdBy)
+        const userForExists = await checkUser.getUserByID(transaction)
+        if (!userForExists) {
+            throw new CustomError.BadRequestError("User not found")
+        }
+
         // Create bucket for storing new user after getting it from view.
         let newUser = await userObj.createUser(transaction)
 
@@ -132,6 +128,13 @@ const updateUser = async (userObj) => {
             throw new CustomError.BadRequestError("User email already exists")
         }
 
+        // Check if updated by exists in db.
+        const checkUser = User.createBlankUserWithID(userObj.updatedBy)
+        const userForExists = await checkUser.getUserByID(transaction)
+        if (!userForExists) {
+            throw new CustomError.BadRequestError("User not found")
+        }
+
         // Create bucket for storing updated user after getting it from view.
         let updateUser = await userObj.updateUser(transaction)
 
@@ -158,6 +161,13 @@ const deleteUser = async (userObj) => {
     const transaction = await db.sequelize.transaction()
 
     try {
+
+        // Check if deleted by exists in db.
+        const checkUser = User.createBlankUserWithID(userObj.deletedBy)
+        const userForExists = await checkUser.getUserByID(transaction)
+        if (!userForExists) {
+            throw new CustomError.BadRequestError("User not found")
+        }
 
         // Create bucket for storing deleted user after getting it from view.
         let deleteUser = await userObj.deleteUser(transaction)
@@ -203,6 +213,43 @@ const login = async (userObj) => {
         // Return error.
         throw error
     }
+}
+
+// Add search queries to  query.
+const addSearchQueries = (params) => {
+
+    // Create bucket for where query.
+    let where = {}
+
+    // First name.
+    if (params.firstName) {
+        where.firstName = {
+            [Op.iLike]: "%" + params.firstName.toLowerCase() + "%"
+        }
+    }
+
+    // Last name.
+    if (params.lastName) {
+        where.lastName = {
+            [Op.iLike]: "%" + params.lastName.toLowerCase() + "%"
+        }
+    }
+
+    // Email.
+    if (params.email) {
+        where.email = {
+            [Op.iLike]: "%" + params.email.toLowerCase() + "%"
+        }
+    }
+
+    // Is Admin.
+    if (params.isAdmin) {
+        where.isAdmin = {
+            [Op.eq]: params.isAdmin
+        }
+    }
+
+    return where
 }
 
 // Export all the functions.
